@@ -36,61 +36,70 @@
     "lint": "eslint --ext .js,.jsx,.ts,.tsx,.mjs ./src/",
     "test": "jest",
     "bundle": "webpack",
-    "bundle:prod": "webpack -p",
+    "bundle:prod": "webpack --mode production",
     "aws:copy": "aws s3 sync ./dist s3://ts-spa-boilerplate",
     "aws:open": "start http://ts-spa-boilerplate.s3-website-eu-west-1.amazonaws.com"
   },
   "dependencies": {},
   "devDependencies": {
-    "@types/jest": "^23.3.1",
-    "@types/react": "^16.4.9",
-    "autoprefixer": "^9.1.0",
+    "@types/express": "^4.16.0",
+    "@types/jest": "^23.3.2",
+    "@types/react": "^16.4.14",
+    "@types/react-dom": "^16.0.7",
+    "@types/react-router-dom": "^4.3.1",
+    "@types/webpack": "^4.4.11",
+    "@types/webpack-dev-middleware": "^2.0.2",
+    "@types/webpack-hot-middleware": "^2.16.4",
+    "autoprefixer": "^9.1.5",
     "bulma": "^0.7.1",
     "clean-webpack-plugin": "^0.1.19",
     "copy-webpack-plugin": "^4.5.2",
     "cross-env": "^5.2.0",
     "css-loader": "^1.0.0",
-    "cssnano": "^4.0.5",
+    "cssnano": "^4.1.0",
     "es6-cachemanager": "0.0.2",
-    "es6-promise": "^4.2.4",
-    "eslint": "^5.3.0",
+    "es6-promise": "^4.2.5",
+    "eslint": "^5.6.0",
     "eslint-config-eser": "^2.0.7",
     "eslint-config-eser-react": "^2.0.7",
-    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-react": "^7.10.0",
+    "eslint-plugin-react": "^7.11.1",
     "express": "^4.16.3",
-    "file-loader": "^1.1.11",
+    "file-loader": "^2.0.0",
     "font-awesome": "^4.7.0",
-    "jest": "^23.4.2",
+    "jest": "^23.6.0",
     "json-loader": "^0.5.7",
-    "mini-css-extract-plugin": "^0.4.1",
+    "mini-css-extract-plugin": "^0.4.2",
     "node-sass": "^4.9.3",
     "path-browser": "^2.2.1",
     "postcss": "^7.0.2",
     "postcss-loader": "^3.0.0",
     "postcss-modules": "^1.3.2",
+    "postcss-nesting": "^6.0.0",
     "prop-types": "^15.6.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "react": "^16.5.1",
+    "react-dom": "^16.5.1",
     "react-eventmanager": "0.0.3",
-    "react-intl": "^2.4.0",
-    "react-markdown": "^3.4.1",
+    "react-intl": "^2.5.0",
+    "react-markdown": "^3.6.0",
     "react-router-dom": "^4.3.1",
     "sass-loader": "^7.1.0",
     "servicemanager": "^0.1.0",
-    "style-loader": "^0.22.1",
-    "ts-jest": "^23.1.3",
-    "ts-loader": "^4.4.2",
+    "style-loader": "^0.23.0",
+    "ts-jest": "^23.1.4",
+    "ts-loader": "^5.1.1",
     "tslib": "^1.9.3",
-    "typescript": "^3.0.1",
+    "typescript": "^3.0.3",
     "typescript-eslint-parser": "^18.0.0",
-    "webpack": "^4.16.5",
-    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack": "^4.19.0",
+    "webpack-async-chunk-names-plugin": "^0.1.1",
+    "webpack-bundle-analyzer": "^3.0.2",
     "webpack-cli": "^3.1.0",
-    "webpack-dev-middleware": "^3.1.3",
+    "webpack-dev-middleware": "^3.3.0",
     "webpack-dotenv-plugin": "^2.1.0",
-    "webpack-hot-middleware": "^2.22.3",
-    "whatwg-fetch": "^2.0.4"
+    "webpack-hot-middleware": "^2.24.0",
+    "webpack-node-externals": "^1.7.2",
+    "whatwg-fetch": "^3.0.0"
   }
 }
